不知名 BUG 所以改變查詢 table 的方法

+++ discuzx/utility/convert/source/d7.2_x2.0/optionvalue.php
@@ -22,19 +22,24 @@ $_sorts = getgpc('_sorts');
 if(empty($_sorts)) {
 	$_sorts = array();
 	// 搜尋相關的 TABLE
-	$query = $db_source->query("SHOW TABLES LIKE '{$config[source][tablepre]}optionvalue%'");
-	while($row = $db_source->fetch_row($query)) {
+	$query = mysql_list_tables($db_source->config['dbname'], $db_source->curlink);
+	$num_rows = mysql_num_rows($query);
+	for ($i = 0; $i < $num_rows; $i++) {
+		$row = mysql_tablename($query, $i);
+
+		if (strpos($row, $db_source->config['tablepre'].'optionvalue') !== 0) continue;
+
 		$tabledump = '';
 		$sortid = 0;
 		// 檢查是否為正確想要搜尋的 TABLE
-		$sortid = str_replace($config[source][tablepre].'optionvalue', '', $row[0]);
-		if ($sortid && $sortid == intval($sortid)) {
+		$sortid = str_replace($db_source->config['tablepre'].'optionvalue', '', $row);
+		if ($sortid && $sortid == intval($sortid) && is_numeric($sortid)) {
 			// 複製來源 TABLE 結構
-			if ($tabledump = _sqldumptablestruct($config[source][tablepre].'optionvalue'.$sortid, $db_source)) {
+			if ($tabledump = _sqldumptablestruct($db_source->config['tablepre'].'optionvalue'.$sortid, $db_source)) {
 				// 刪除已存在的目標 TABLE
 				$db_target->query("DROP TABLE IF EXISTS ".$table_target.$sortid);
 				// 轉換 TABLE 表的建立語法 將來源 TABLE 名稱取代為目標 TABLE 名稱
-				$tabledump = str_replace('CREATE TABLE `'.$config[source][tablepre].'optionvalue'.$sortid.'`', 'CREATE TABLE `'.$config[target][tablepre].'forum_optionvalue'.$sortid.'`', $tabledump);
+				$tabledump = str_replace('CREATE TABLE `'.$db_source->config['tablepre'].'optionvalue'.$sortid.'`', 'CREATE TABLE `'.$db_target->config['tablepre'].'forum_optionvalue'.$sortid.'`', $tabledump);
 				$db_target->query($tabledump);
 			}
 			$_sorts[] = $sortid;
